+++ src/app/_components/layout/HomeHeader.tsx
@@ -1,6 +1,6 @@
-import AlarmButton from 'app/inapp/_components/layout/AlarmButton';
-import FavoriteCitiesDropdown from 'app/inapp/_components/layout/FavoriteCitiesDropdown';
-import SearchButton from 'app/inapp/_components/layout/SearchButton';
+import AlarmButton from 'app/_components/layout/AlarmButton';
+import FavoriteCitiesDropdown from 'app/_components/layout/FavoriteCitiesDropdown';
+import SearchButton from 'app/_components/layout/SearchButton';
 
 const HomeHeader = () => {
   return (
