+++ src/app/_components/product/ProductInfo.tsx
@@ -1,7 +1,7 @@
-import ProductPrice from 'app/inapp/_components/product/ProductPrice';
-import ProductRegistrationMeta from 'app/inapp/_components/product/ProductRegistrationMeta';
-import ProductTitle from 'app/inapp/_components/product/ProductTitle';
-import ProductUserReaction from 'app/inapp/_components/product/ProductUserReaction';
+import ProductPrice from 'app/_components/product/ProductPrice';
+import ProductRegistrationMeta from 'app/_components/product/ProductRegistrationMeta';
+import ProductTitle from 'app/_components/product/ProductTitle';
+import ProductUserReaction from 'app/_components/product/ProductUserReaction';
 
 interface Props {
   title: string;
