+++ src/app/_components/product/ProductRegistrationMeta.tsx
