+++ src/app/favorite/_components/search/SearchFavoriteCitiesSection.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useState } from 'react';
-import SearchedCitiesList from 'app/inapp/favorite/_components/search/SearchedCitiesList';
-import SearchFavoriteCitiesInput from 'app/inapp/favorite/_components/search/SearchFavoriteCitiesInput';
-import SelectedFavoriteCitiesList from 'app/inapp/favorite/_components/search/SelectedFavoriteCitiesList';
+import SearchedCitiesList from 'app/favorite/_components/search/SearchedCitiesList';
+import SearchFavoriteCitiesInput from 'app/favorite/_components/search/SearchFavoriteCitiesInput';
+import SelectedFavoriteCitiesList from 'app/favorite/_components/search/SelectedFavoriteCitiesList';
 
 const SearchFavoriteCitiesSection = () => {
   const [isInputFocused, setIsInputFocused] = useState(false);
