+++ src/app/favorite/_components/search/SearchInputIcon.tsx
