+++ src/app/favorite/_components/search/SelectedFavoriteCitiesList.tsx
