+++ src/app/favorite/page.tsx
@@ -1,0 +1,15 @@
+import CitiesSettingDescription from 'app/favorite/_components/description/CitiesSettingDescription';
+import CountryLabel from 'app/favorite/_components/description/CountryLabel';
+import SearchFavoriteCitiesSection from 'app/favorite/_components/search/SearchFavoriteCitiesSection';
+
+const Page = () => {
+  return (
+    <>
+      <CountryLabel />
+      <CitiesSettingDescription />
+      <SearchFavoriteCitiesSection />
+    </>
+  );
+};
+
+export default Page;
