+++ src/app/register/_components/registerForm/RegisterImageForm.tsx
@@ -14,7 +14,7 @@ import {
   horizontalListSortingStrategy,
   SortableContext,
 } from '@dnd-kit/sortable';
-import SortableImageItem from 'app/inapp/register/_components/registerForm/SortableImageItem';
+import SortableImageItem from 'app/register/_components/registerForm/SortableImageItem';
 
 const MAX_IMAGES_COUNT = 10;
