+++ src/app/register/_components/registerForm/category/CategoryList.tsx
@@ -1,5 +1,5 @@
 import { UseFormSetValue } from 'react-hook-form';
-import { FormDataType } from 'app/inapp/register/_type/registerType';
+import { FormDataType } from 'app/register/_type/registerType';
 import { categoryTestDataType } from 'app/testDatas/categoryTestData';
 
 type Props = {
