+++ src/app/register/_components/registerForm/category/CategoryModal.tsx
@@ -1,7 +1,7 @@
 import { UseFormSetValue } from 'react-hook-form';
-import CategoryList from 'app/inapp/register/_components/registerForm/category/CategoryList';
-import RegisterCategoryHeader from 'app/inapp/register/_components/registerForm/category/RegisterCategoryHeader';
-import { FormDataType } from 'app/inapp/register/_type/registerType';
+import CategoryList from 'app/register/_components/registerForm/category/CategoryList';
+import RegisterCategoryHeader from 'app/register/_components/registerForm/category/RegisterCategoryHeader';
+import { FormDataType } from 'app/register/_type/registerType';
 import { categoryTestData } from 'app/testDatas/categoryTestData';
 
 type Props = {
