+++ src/app/register/_type/registerType.ts
